requirements/common.txt
Django==2.2.1
django-reversion==3.0.4
